requirements.txt
cffi==1.14.4
cryptography==3.2.1
Django==2.2
django-admin-list-filter-dropdown==1.0.3
django-cors-headers==3.6.0
django-filter==2.4.0
djangorestframework==3.12.2
flake8==3.8.4
mccabe==0.6.1
migrate==0.3.8
psycopg2-binary==2.8.6
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
PyJWT==2.0.0
pytz==2020.5
six==1.15.0
sqlparse==0.4.1
